+++ tests/testthat/test_io.R
@@ -73,9 +73,71 @@ with(test_data, {
     expect_true(all(locus_attrs == locus_attrs_test))
   })
 
+  test_that("load_csv parses CSV files with unknown columns", {
+    # For cases where our CSV file has none of the recognized column names, it
+    # should still work but just give us generic rownames and a
+    # warning.
+    fp <- tempfile()
+    data_expected <- data.frame(
+      Vec1 = c("A", "D"),
+      Vec2 = c("B", "E"),
+      Vec3 = c("C", "F"),
+      stringsAsFactors = FALSE,
+      row.names = c("entry1", "entry2"))
+    cat("Vec1,Vec2,Vec3\nA,B,C\nD,E,F\n", file = fp)
+    expect_warning(data <- load_csv(fp), "no recognized columns for entry id")
+    file.remove(fp)
+    expect_equal(data, data_expected)
+  })
+
+  test_that("load_csv uses row names given", {
+    # If we give a row.names argument it should pass that on to read.table and
+    # shouldn't automatically generate row names
+    fp <- tempfile()
+    data_expected <- data.frame(
+      Vec1 = c("A", "D"),
+      Vec2 = c("B", "E"),
+      Vec3 = c("C", "F"),
+      stringsAsFactors = FALSE,
+      row.names = NULL)
+    cat("Vec1,Vec2,Vec3\nA,B,C\nD,E,F\n", file = fp)
+    data <- load_csv(fp, row.names = NULL)
+    file.remove(fp)
+    expect_equal(data, data_expected)
+  })
+
 
   test_that("save_csv saves CSV files", {
     fp <- tempfile()
+    save_csv(locus_attrs, fp)
+    locus_attrs_test <- load_locus_attrs(fp)
+    file.remove(fp)
+    expect_equal(nrow(locus_attrs_test), nrow(locus_attrs))
+    expect_equal(ncol(locus_attrs_test), ncol(locus_attrs))
+    expect_equal(locus_attrs_cols, colnames(locus_attrs_test))
+    expect_equal(c("A", "B", "1", "2"), rownames(locus_attrs_test))
+    expect_true(all(locus_attrs == locus_attrs_test))
+  })
+
+  test_that("save_csv saves CSV files with unknown columns", {
+    # save_csv shouldn't care about the columns but we'll make sure here.
+    fp <- tempfile()
+    data_expected <- data.frame(
+      Vec1 = c("A", "D"),
+      Vec2 = c("B", "E"),
+      Vec3 = c("C", "F"),
+      stringsAsFactors = FALSE)
+    save_csv(data_expected, fp)
+    expect_warning(data <- load_csv(fp))
+    file.remove(fp)
+    rownames(data_expected) <- c("entry1", "entry2")
+    expect_equal(data, data_expected)
+  })
+
+  test_that("save_csv makes parent dirs when saving", {
+    # the parent directories here don't yet exist, so for this to work,
+    # save_csv will need to create them
+    fp <- file.path(tempfile(), basename(tempfile()))
     save_csv(locus_attrs, fp)
     locus_attrs_test <- load_locus_attrs(fp)
     file.remove(fp)